refactor: move navbar from components to root
