+++ src/app/routes/contact.tsx
@@ -4,7 +4,6 @@ import classNames from 'classnames';
 import { useEffect, useState, useRef } from 'react';
 import { z } from 'zod';
 import { FormErrorMessage } from '~/components/FormErrorMessage';
-import NavBar from '~/components/NavBar';
 import { H1 } from '~/components/ui/H1';
 import { Button } from '~/components/ui/button';
 import { Input } from '~/components/ui/input';
@@ -243,7 +242,6 @@ export default function Contact() {
 
     return (
         <div>
-            <NavBar login={data.login} role={data.role} />
             <div className='md:flex md:justify-center mx-4 md:mx-0'>
                 <H1 className='my-4 md:w-3/5'>Contact the Student Council</H1>
             </div>
